feat: add gap of each node
